Move themeActionBar to AndroidX theme utils, since it's an appcompat class
