+++ ui/src/main/java/com/nextcloud/android/common/ui/theme/utils/MaterialViewThemeUtils.kt
@@ -27,15 +27,7 @@ import android.content.Context
 import android.content.res.ColorStateList
 import android.graphics.Color
 import android.graphics.PorterDuff
-import android.graphics.drawable.ColorDrawable
-import android.graphics.drawable.Drawable
-import android.text.Spannable
-import android.text.SpannableString
-import android.text.style.ForegroundColorSpan
-import androidx.annotation.ColorInt
-import androidx.appcompat.app.ActionBar
 import androidx.core.content.ContextCompat
-import androidx.core.graphics.drawable.DrawableCompat
 import com.google.android.material.appbar.MaterialToolbar
 import com.google.android.material.button.MaterialButton
 import com.google.android.material.card.MaterialCardView
@@ -206,32 +198,6 @@ class MaterialViewThemeUtils @Inject constructor(schemes: MaterialSchemes, priva
                 )
             )
         }
-    }
-
-    fun themeActionBar(context: Context, actionBar: ActionBar, title: String, backArrow: Drawable) {
-        withScheme(context) { scheme ->
-            setColoredTitle(actionBar, title, scheme.onSurface)
-            actionBar.setBackgroundDrawable(ColorDrawable(scheme.surface))
-
-            val wrap = DrawableCompat.wrap(backArrow)
-            wrap.setColorFilter(scheme.onSurface, PorterDuff.Mode.SRC_ATOP)
-            actionBar.setHomeAsUpIndicator(wrap)
-        }
-    }
-
-    private fun setColoredTitle(
-        actionBar: ActionBar,
-        title: String,
-        @ColorInt color: Int
-    ) {
-        val text: Spannable = SpannableString(title)
-        text.setSpan(
-            ForegroundColorSpan(color),
-            0,
-            text.length,
-            Spannable.SPAN_INCLUSIVE_INCLUSIVE
-        )
-        actionBar.title = text
     }
 
     fun themeToolbar(toolbar: MaterialToolbar) {
